requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.2
fonttools==4.39.2
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
html2text==2020.1.16
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
multidict==6.0.4
mysql-connector-python==8.0.33
mysql-connector-python-rf==2.2.2
numpy==1.24.2
openai==0.27.2
orjson==3.8.8
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.2
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
soupsieve==2.4
starlette==0.27.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.2
tqdm==4.65.0
typing_extensions==4.6.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
